Revert "Changing how completed is marked"
