+++ tests/conversations/test_ops.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta, timezone
 
 from pytz import UTC
-from conversations.ops import add_message_to_conversation, update_completed_conversation, update_suspended_conversation
+from conversations.ops import add_message_to_conversation, update_completed_conversation
 from datatypes import Conversation, ClassifiedMessage, CalendarClassification
 
 
@@ -33,49 +33,8 @@ def test_update_marks_only_old_conversations_as_completed():
     older_conv = Conversation(last_updated=current_time - timedelta(seconds=40))
     conversations = [current_conv, slightly_old_conv, older_conv]
 
-    updated_conversations = update_suspended_conversation(conversations, 30, current_time)
+    updated_conversations = update_completed_conversation(conversations, 30, current_time)
 
-    assert [conv.suspended for conv in updated_conversations] == [False, False, True]
+    assert [conv.completed for conv in updated_conversations] == [False, False, True]
 
 
-def test_update_completed_conversation():
-    # Create some sample data
-    past_datetime = datetime.now() - timedelta(days=1)
-    future_datetime = datetime.now() + timedelta(days=1)
-
-    conversation1 = Conversation(
-        event_datetime=past_datetime,
-        lines=[],
-        users=set(),
-        last_updated=None,
-        suspended=False,
-        completed=False
-    )
-
-    conversation2 = Conversation(
-        event_datetime=future_datetime,
-        lines=[],
-        users=set(),
-        last_updated=None,
-        suspended=False,
-        completed=False
-    )
-
-    conversation3 = Conversation(
-        event_datetime=None,
-        lines=[],
-        users=set(),
-        last_updated=None,
-        suspended=False,
-        completed=False
-    )
-
-    conversations = [conversation1, conversation2, conversation3]
-
-    # Call the function with a current time in between past and future datetime
-    updated_conversations = update_completed_conversation(conversations, datetime.now())
-
-    # Check that the completed status is correctly set
-    assert updated_conversations[0].completed == True, "Conversation 1 should be marked as completed"
-    assert updated_conversations[1].completed == False, "Conversation 2 should not be marked as completed"
-    assert updated_conversations[1].completed == False, "Conversation 2 should remain unchanged"
